Enforce that INSERT...RETURNING preserves the order of multi rows
It has been a fact that the RETURNING clause on an INSERT will return
multiple rows with the same order as multiple VALUES have been fed.

eg: INSERT INTO tbl1(code) VALUES ('abc'), ('def'), ('agh')
           RETURNING id, code;

is expected to yield:
   id | code
  -----------
    1 | abc
    2 | def
    3 | agh

Clarify that in the documentation, and also write a test case that will
prevent us from breaking the rule in the future.